Creating the basics of the personal deck page
I have setup the personal deck page's basic layout. The deck builder portion of the page will be a table that shows the current deck, and it's cards on the left side, then the list of all cards on the right side.

The div with the id of newDeckSetup will be what appears when a user creates a new deck before they get to the deck builder portion

The deckList div will be the section that lists all of the decks the user owns