Created the main card list portion of the deck building page
I need to make it so that the cards can be added to the deck, and decks can be created and loaded and all of that fun stuff. There is also currently a bug that makes it so that when you click on a card, it will no longer go back to the :hover code, so that's a thing.